restructred program analysis tools
I'm suggesting to leave "Code Analysis and Linters" section only for tools that work on OCaml. The rest of the tools I merged with the "Formal Software Verification" section, and added few more libraries, like frama-c, cil, and bap. I also renamed the section to make it more general, by removing Formal, and adding Program Analysis. There are actually a plethora of program analysis tools, that are not mentioned yet, although I'm not sure, should we mention tools that are proprietary or no longer supported (in a dormant state). 

Additionally, I fixed the alphabet ordering for regular expressions, and made the a subsection (not a section)