Parallelism : Sklml
 Sklml, the OCaml parallel skeleton system, is a functional parallel skeleton compiler and programming system for OCaml programs.

Writing parallel programs is not easy, and debugging them is usually a nightmare. To cope with these difficulties, the skeleton programming approach uses a set of predefined patterns for parallel computations. The skeletons are higher order functional templates that describe the program underlying parallelism.

Sklml is a new framework for parallel programming that embeds an innovative compositional skeleton algebra into the OCaml language. Thanks to its skeleton algebra, Sklml provides two evaluation regimes to programs: a regular sequential evaluation (merely used for prototyping and debugging) and a parallel evaluation obtained via a recompilation of the same source program in parallel mode.

Sklml was specifically designed to prove that the sequential and parallel evaluation regimes coincide.